recommendation-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymysql==1.1.1
sqlalchemy==2.0.36
pandas==2.2.3
scikit-learn==1.5.2
redis==5.2.0
pyjwt==2.9.0
python-dotenv==1.0.1
cryptography==44.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
